Feat: 엔드포인트 인가 처리 (#87)

+++ src/main/java/edu/handong/csee/histudy/controller/AdminController.java
@@ -1,9 +1,11 @@
 package edu.handong.csee.histudy.controller;
 
+import edu.handong.csee.histudy.domain.Role;
 import edu.handong.csee.histudy.dto.TeamDto;
 import edu.handong.csee.histudy.dto.TeamIdDto;
 import edu.handong.csee.histudy.dto.TeamReportDto;
 import edu.handong.csee.histudy.dto.UserDto;
+import edu.handong.csee.histudy.exception.ForbiddenException;
 import edu.handong.csee.histudy.service.TeamService;
 import edu.handong.csee.histudy.service.UserService;
 import io.jsonwebtoken.Claims;
@@ -26,16 +28,26 @@ public class AdminController {
     private final TeamService teamService;
     private final UserService userService;
 
-    @Operation(summary = "매칭된 그룹 목록 조회")
+    @Operation(summary = "그룹별 활동 조회")
     @GetMapping(value = "/manageGroup")
     public ResponseEntity<List<TeamDto>> getTeams(@RequestAttribute Claims claims) {
-        return ResponseEntity.ok(teamService.getTeams(claims.getSubject()));
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            String email = claims.getSubject();
+            return ResponseEntity.ok(teamService.getTeams(email));
+        }
+        throw new ForbiddenException();
     }
 
+    @Deprecated
     @Operation(summary = "그룹 삭제")
     @DeleteMapping("/group")
-    public ResponseEntity<Integer> deleteTeam(@RequestBody TeamIdDto dto, @RequestAttribute Claims claims) {
-        return ResponseEntity.ok(teamService.deleteTeam(dto, claims.getSubject()));
+    public ResponseEntity<Integer> deleteTeam(
+            @RequestBody TeamIdDto dto,
+            @RequestAttribute Claims claims) {
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return ResponseEntity.ok(teamService.deleteTeam(dto, claims.getSubject()));
+        }
+        throw new ForbiddenException();
     }
 
     @Operation(summary = "특정 그룹 보고서 조회")
@@ -44,36 +56,58 @@ public class AdminController {
             @Parameter(description = "그룹 아이디", required = true)
             @PathVariable(name = "id") long id,
             @RequestAttribute Claims claims) {
-        return ResponseEntity.ok(teamService.getTeamReports(id, claims.getSubject()));
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return ResponseEntity.ok(teamService.getTeamReports(id, claims.getSubject()));
+        }
+        throw new ForbiddenException();
     }
 
     @Operation(summary = "신청한 유저 목록 조회")
     @GetMapping("/allUsers")
-    public ResponseEntity<List<UserDto.UserInfo>> getAppliedUsers() {
-        return ResponseEntity.ok(userService.getAppliedUsers());
+    public ResponseEntity<List<UserDto.UserInfo>> getAppliedUsers(@RequestAttribute Claims claims) {
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return ResponseEntity.ok(userService.getAppliedUsers());
+        }
+        throw new ForbiddenException();
     }
 
     @Operation(summary = "그룹 매칭")
     @PostMapping("/team-match")
-    public ResponseEntity<TeamDto.MatchResults> matchTeam() {
-        return ResponseEntity.ok(teamService.matchTeam());
+    public ResponseEntity<TeamDto.MatchResults> matchTeam(@RequestAttribute Claims claims) {
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return ResponseEntity.ok(teamService.matchTeam());
+        }
+        throw new ForbiddenException();
     }
 
     @Operation(summary = "매칭되지 않은 유저 목록 조회")
     @GetMapping("/unmatched-users")
-    public ResponseEntity<List<UserDto.UserInfo>> getUnmatchedUsers() {
-        return ResponseEntity.ok(userService.getUnmatchedUsers());
+    public ResponseEntity<List<UserDto.UserInfo>> getUnmatchedUsers(@RequestAttribute Claims claims) {
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return ResponseEntity.ok(userService.getUnmatchedUsers());
+        }
+        throw new ForbiddenException();
     }
 
     @Operation(summary = "특정 유저 지원폼 삭제")
     @DeleteMapping("/form")
-    public UserDto.UserInfo deleteForm(@RequestParam String sid) {
-        return userService.deleteUserForm(sid);
+    public UserDto.UserInfo deleteForm(
+            @RequestParam String sid,
+            @RequestAttribute Claims claims) {
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return userService.deleteUserForm(sid);
+        }
+        throw new ForbiddenException();
     }
 
     @Operation(summary = "유저 정보 수정")
     @PostMapping("/edit-user")
-    public UserDto.UserInfo editUser(@RequestBody UserDto.UserEdit dto) {
-        return userService.editUser(dto);
+    public UserDto.UserInfo editUser(
+            @RequestBody UserDto.UserEdit dto,
+            @RequestAttribute Claims claims) {
+        if (Role.isAuthorized(claims, Role.ADMIN)) {
+            return userService.editUser(dto);
+        }
+        throw new ForbiddenException();
     }
 }
